driver: Move bucket name generation to a separate module
Instead of hardcoding how the bucket name is generated, and only
allowing the user to set the number of buckets, move the whole
generation into a separate module, and use the standard way of
specifying generators to construct bucket names.

This makes the code not only more flexible, but simpler too.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>